pyproject.toml
[build-system]
requires = ['hatchling>=1.21']
build-backend = 'hatchling.build'

[project]
name = 'jpl.labcas.backend'
dynamic = ['version']
description = 'Python implementation of the LabCAS data access API with flexible security system.'
keywords = ['data', 'LabCAS', 'cancer', 'backend', 'jpl', 'nist', 'edrn', 'nasa']
requires-python = '>=3.13,<3.15'
license = {file = 'LICENSE.md'}
authors = [
    {name = 'Sean Kelly', email = 'kelly@seankelly.biz'}
]
readme = 'README.md'
dependencies = [
    'fastapi>=0.115.0',
    'uvicorn[standard]>=0.30.0',
    'httpx>=0.27.0',
    'boto3>=1.34.0',
    'python-jose[cryptography]>=3.3.0',
    'redis>=5.0.0',
    'ldap3>=2.9.0',
    'python-dotenv>=1.0.0',
    'pydantic>=2.6.0',
    'pydantic-settings>=2.5.2',
    'python-multipart~=0.0.20',
    'tqdm~=4.66.0'
]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Healthcare Industry',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
]

[project.optional-dependencies]
dev = [ 
    'pytest>=8.2.0',
    'pytest-asyncio>=0.23.0',
    'anyio>=4.0.0'
]

[project.urls]
homepage = 'https://github.com/jpl-labcas/backend'
issues = 'https://github.com/jpl-labcas/backend/issues'

[project.scripts]
labcas-backend = 'jpl.labcas.backend.main:cli'
migrate-solr = 'jpl.labcas.backend.search.migration:main'

[tool.pytest.ini_options]
minversion = '8.0'
addopts = '-q'
testpaths = [
    'tests'
]

[tool.hatch.version]
path = 'src/jpl/labcas/backend/VERSION.txt'
pattern = '(?P<version>.+)'

[tool.hatch.build.targets.wheel]
packages = ['src/jpl']
